feat: do not proceed if temporary handler exists

+++ index.js
@@ -157,24 +157,34 @@ module.exports = function getPlugin(S) {
       const absolutePath = path.join(pathSource, relativeWrapperPath);
 
       // 0. Check if file exist in path
-      return fs.move(serverlessHandlerPath, savedHandlerPath, { overwrite: true })
-        .then(() => {
-          // 2. Generate wrapped handler code
-          return codeTemplate({
-            orig_handler_path: `./${savedHandlerFilename}`,
-            wrapper_path: relativeWrapperPath.replace(/\\/g, '/'), // Support Windows env
-            handler_name: handlerFunction,
+      return fs.pathExists(savedHandlerPath).then(tmpHandlerExists => {
+        if (isLocalRun && tmpHandlerExists) {
+          throw new Error(
+            `Cannot wrap lambda: Temporary wrapper file found at ${savedHandlerPath}\n` +
+            'It\'s likely that previous run crashed, and left broken state.\n' +
+            'If it\'s the case, then replacing lambda handler file with content of ' +
+            `${savedHandlerFilename}, and removing the ${savedHandlerFilename} should fix the issue.`
+          );
+        }
+        return fs.move(serverlessHandlerPath, savedHandlerPath, { overwrite: true })
+          .then(() => {
+            // 2. Generate wrapped handler code
+            return codeTemplate({
+              orig_handler_path: `./${savedHandlerFilename}`,
+              wrapper_path: relativeWrapperPath.replace(/\\/g, '/'), // Support Windows env
+              handler_name: handlerFunction,
+            });
+          })
+          .then(code => {
+            // 3. Write code to wrapped handler
+            return fs.outputFile(wrappedServerlessHandlerPath, code);
+          })
+          .then(() => {
+            // 4. Resolve the event
+            SCli.log(`Wrapping ${handler} with ${absolutePath}`);
+            return evt;
           });
-        })
-        .then(code => {
-          // 3. Write code to wrapped handler
-          return fs.outputFile(wrappedServerlessHandlerPath, code);
-        })
-        .then(() => {
-          // 4. Resolve the event
-          SCli.log(`Wrapping ${handler} with ${absolutePath}`);
-          return evt;
-        });
+      });
     }
 
     // Information about the serverless framework version of the handler
